@@ -1,5 +1,8 @@
+import 'package:fair/fair.dart';
+
+@FairBinding()
 class MealsListData {
-  MealsListData({
+  const MealsListData({
     this.imagePath = '',
     this.titleTxt = '',
     this.startColor = '',
@@ -8,14 +11,54 @@ class MealsListData {
     this.kacl = 0,
   });
 
-  String imagePath;
-  String titleTxt;
-  String startColor;
-  String endColor;
-  List<String> meals;
-  int kacl;
+  final String imagePath;
+  final String titleTxt;
+  final String startColor;
+  final String endColor;
+  final List<String>? meals;
+  final int kacl;
 
-  static List<MealsListData> tabIconsList = <MealsListData>[
+  static const icon1 = MealsListData(
+    imagePath: 'assets/fitness_app/breakfast.png',
+    titleTxt: 'Breakfast',
+    kacl: 525,
+    meals: <String>['Bread,', 'Peanut butter,', 'Apple'],
+    startColor: '#FA7D82',
+    endColor: '#FFB295',
+  );
+
+  static const icon2 = MealsListData(
+    imagePath: 'assets/fitness_app/lunch.png',
+    titleTxt: 'Lunch',
+    kacl: 602,
+    meals: <String>['Salmon,', 'Mixed veggies,', 'Avocado'],
+    startColor: '#738AE6',
+    endColor: '#5C5EDD',
+  );
+
+  static const icon3 = MealsListData(
+    imagePath: 'assets/fitness_app/snack.png',
+    titleTxt: 'Snack',
+    kacl: 0,
+    meals: <String>['Recommend:', '800 kcal'],
+    startColor: '#FE95B6',
+    endColor: '#FF5287',
+  );
+
+  static const icon4 = MealsListData(
+    imagePath: 'assets/fitness_app/dinner.png',
+    titleTxt: 'Dinner',
+    kacl: 0,
+    meals: <String>['Recommend:', '703 kcal'],
+    startColor: '#6F72CA',
+    endColor: '#1E1466',
+  );
+
+  static List<MealsListData> getList() {
+    return tabIconsList;
+  }
+
+  static const List<MealsListData> tabIconsList = <MealsListData>[
     MealsListData(
       imagePath: 'assets/fitness_app/breakfast.png',
       titleTxt: 'Breakfast',
