@@ -1,4 +1,5 @@
 import 'package:best_flutter_ui_templates/design_course/home_design_course.dart';
+import 'package:best_flutter_ui_templates/fair/page/my_diary_page.dart';
 import 'package:best_flutter_ui_templates/fitness_app/fitness_app_home_screen.dart';
 import 'package:best_flutter_ui_templates/hotel_booking/hotel_home_screen.dart';
 import 'package:flutter/widgets.dart';
@@ -9,8 +10,8 @@ class HomeList {
     this.imagePath = '',
   });
 
-  Widget navigateScreen;
-  String imagePath;
+  Widget? navigateScreen;
+  String? imagePath;
 
   static List<HomeList> homeList = [
     HomeList(
@@ -20,6 +21,7 @@ class HomeList {
     HomeList(
       imagePath: 'assets/fitness_app/fitness_app.png',
       navigateScreen: FitnessAppHomeScreen(),
+      // navigateScreen: MyDiaryPage()
     ),
     HomeList(
       imagePath: 'assets/design_course/design_course.png',
