@@ -21,6 +21,7 @@ class AppTheme {
   static const Color chipBackground = Color(0xFFEEF1F3);
   static const Color spacer = Color(0xFFF2F2F2);
   static const String fontName = 'WorkSans';
+  static const Color greyWithOpacity = Color(0xD99E9E9E);
 
   static const TextTheme textTheme = TextTheme(
     headline4: display1,
@@ -95,6 +96,4 @@ class AppTheme {
     letterSpacing: 0.2,
     color: lightText, // was lightText
   );
-
-  static Color greyWithOpacity = Colors.grey.withOpacity(0.8);
 }
