@@ -10,14 +10,14 @@ import 'src/generated.fair.dart' as p;
 
 @FairBinding(packages: [
   'package:font_awesome_flutter/font_awesome_flutter.dart',
-  'package:smooth_star_rating/smooth_star_rating.dart'
+  'package:flutter_rating_bar/flutter_rating_bar.dart',
 ])
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
   await SystemChrome.setPreferredOrientations(<DeviceOrientation>[
     DeviceOrientation.portraitUp,
     DeviceOrientation.portraitDown
-  ]).then((_) => runApp(MyApp()));
+  ]).then((_) => FairApp.runApplication(MyApp()));
 }
 
 class MyApp extends StatelessWidget {
@@ -33,9 +33,9 @@ class MyApp extends StatelessWidget {
       systemNavigationBarIconBrightness: Brightness.dark,
     ));
     return FairApp(
-      child:  MaterialApp(
+      child: MaterialApp(
         title: 'Flutter UI',
-        debugShowCheckedModeBanner: false,
+        debugShowCheckedModeBanner: true,
         theme: ThemeData(
           primarySwatch: Colors.blue,
           textTheme: AppTheme.textTheme,
